simplify rabbitmq component

+++ app/action/service_utils.py
@@ -71,7 +71,12 @@ def synchronize_data(mqttClient):
             logging.warning(sendData)
             try:
                 mqttClient.publish("stat/V3/" + result.deviceId +"/OEEDATA",json.dumps(sendData))
-                RabbitMQ.getInstance().send_msg(json.dumps(sendData))
+                
+                state = RabbitMQ.getInstance().send_msgV2(json.dumps(sendData))
+                if not state:
+                    logging.error("Error in sending data to RabbitMQ")
+                    raise Exception("Error in sending data to RabbitMQ")
+                
                 # logging.error("stat/V3/" + result.deviceId +"/OEEDATA")
                 db.session.query(UnsyncedMachineData).filter_by(timestamp=result.timestamp).delete()
                 db.session.commit()
